Begin extracting layout classes to separate, acyclic modules
